Airlines Delay Analysis
The aim of this project was to provide insights by showing the potential delays the new airline may encounter in different airports. The objective of this project was to build self-service dynamic and interactive dashboard that shows the performance of airport in term of the metrics of managing flight delay.
This analysis project is based on a dataset that is contained inside one of the sheet in the Excel file of the project. This dataset includes more than 4200 records of internal flight flights in the USA.