Add files via upload
The aim of this project was to provide insights by showing the potential delays the new airline may encounter in different airports. The objective of this project was to build self-service dynamic and interactive dashboard that shows the performance of airport in term of the metrics of managing flight delays.